imovirtual/events/gcs_daily_incremental_load/requirements.txt
pandas
google-cloud-storage
google-cloud-bigquery
google-cloud-exceptions
google-oauth2
python-dotenv